separate msg input to own component
