+++ backoffice/src/components/loby/index.js
@@ -1,7 +1,5 @@
 import { MetaComponent } from '@rebelstack-io/metaflux';
-import { instanceElement, parsemkey } from '../../utils';
-import sendIcon from '../../assets/icons/paper-plane-solid.svg';
-import shoppingIcon from '../../assets/icons/cart-plus-solid.svg';
+import { instanceElement } from '../../utils';
 import '../patchprofile';
 import '../editchannel';
 import '../verificationpopup';
@@ -9,6 +7,7 @@ import '../shoppingcart';
 import './index.css';
 import './msgarea';
 import './msgheader';
+import './msginput';
 
 class Loby extends MetaComponent {
 	/**
@@ -16,23 +15,6 @@ class Loby extends MetaComponent {
 	 */
 	constructor () {
 		super(global.storage);
-	}
-	addListeners() {
-		const inputButton = this.querySelector('#input-button');
-		inputButton.addEventListener('click', () => {
-			const input = document.querySelector('.msg-input > input');
-			this.sendMessage(input);
-		});
-		/*
-		this.querySelector('#settings').addEventListener('click', () => {
-			this.toggleSetting();
-		});*/
-		document.querySelector('.shopping-icon').addEventListener('click', () => {
-			document.querySelector('#shopping-popup-container').classList.toggle('hide');
-		});
-		document.querySelector('#close-shopping').addEventListener('click', ()=> {
-			document.querySelector('#shopping-popup-container').classList.toggle('hide');
-		})
 	}
 	// eslint-disable-next-line class-method-use-this
 	render () {
@@ -64,49 +46,8 @@ class Loby extends MetaComponent {
 	createMsgArea (box) {
 		const msgHeader = instanceElement('msg-header');
 		const msgBodyContainer = instanceElement('msg-area');
-		const msgInput = instanceElement('div', ['msg-input']);
-		this.createInputs(msgInput);
+		const msgInput = instanceElement('msg-input');
 		box.append(msgHeader, msgBodyContainer, msgInput);
-	}
-	/**
-	 * @description create the inputs for the msg area
-	 * @param {HTMLElement} box 
-	 */
-	createInputs (box) {
-		const not = undefined;
-		const input = instanceElement(
-			'input', ['bottom-input'],
-			 not, not,
-			 [{type: 'text'}, {placeholder: 'Enter your message'}]
-		);
-		const inputButton = instanceElement(
-			'div', ['icon', 'send-message-icon'],
-			'input-button',
-			`<img src="${sendIcon}"></img>`,
-			[{type: 'text'}, {placeholder: 'Enter your message'}]
-		);
-		const shopButtom = instanceElement (
-			'div', ['icon', 'shopping-icon'],
-			false,`
-			<img src="${shoppingIcon}"> </img>
-			`
-		)
-		input.addEventListener('keydown', (e) => {
-			if (e.key === 'Enter') {
-				this.sendMessage(input);
-			}
-		});
-		box.append(input, inputButton, shopButtom);
-	}
-	/**
-	 * @description dispatch the send message action
-	 * @param {HTMLElement} input 
-	 */
-	sendMessage (input) {
-		if (input.value !== '') {
-			this.storage.dispatch({ type: 'SEND-MESSAGE', data: input.value, msgType: 'AA' })
-			input.value = '';
-		}
 	}
 	
 	handleStoreEvents () {
