+++ backoffice/src/components/loby/msginput/index.css
@@ -1,0 +1,4 @@
+msg-input {
+    display: grid;
+}
+
